Implemented MorseTree class
Constructor builds takes in a file stream that codes for the morse tree.
Two maps are populated using the data. One map uses the codes as keys
and the other map uses characters as keys.

Encode takes in an uncoded message and outputs a coded version. Throws
an error if an invalid character was used.

Decode takes in a coded message and outputs an uncoded version. Throws
an error if an invalid code segment was used.